Drop the float cast on pi in Camera3D
The cast here never kept the arithmetic in single precision - dividing by a double promotes it straight back.
All it preserved was pi rounded to float first, so dropping it makes the constant more accurate rather than
less. The outdoor half angle moves by 6e-08 and the indoor one does not move at all, since it comes out bit
identical at a 60 degree FOV.

The two casts in Indoor.cpp and the two in Outdoor.cpp stay. Those do pick the arithmetic type, and the
Indoor ones feed a float that comes off grng and places monsters.

+++ src/Engine/Graphics/Camera.cpp
@@ -128,7 +128,7 @@ void Camera3D::CreateViewMatrixAndProjectionScale() {
     ViewPlaneDistPixels = (double)pViewport.w * 0.5 / halfFovTan;
 
     // calculate vertical FOV in degrees for GL rendering
-    fov_y_deg = (180.0 / static_cast<float>(M_PI)) * 2.0 * std::atan((pViewport.h / 2.0) / pCamera3D->ViewPlaneDistPixels);
+    fov_y_deg = (180.0 / M_PI) * 2.0 * std::atan((pViewport.h / 2.0) / pCamera3D->ViewPlaneDistPixels);
 
     screenCenterX = (double)pViewport.center().x;
     screenCenterY = (double)pViewport.center().y - pViewport.y;
@@ -138,11 +138,11 @@ void Camera3D::CreateViewMatrixAndProjectionScale() {
 
 //----- (004374E8) --------------------------------------------------------
 void Camera3D::BuildViewFrustum() {
-    float HalfAngleX = (static_cast<float>(M_PI) / 2.0) - (odm_fov_rad / 2.0);
-    float HalfAngleY = (static_cast<float>(M_PI) / 2.0) - (std::atan((pViewport.h / 2.0) / pCamera3D->ViewPlaneDistPixels));
+    float HalfAngleX = (M_PI / 2.0) - (odm_fov_rad / 2.0);
+    float HalfAngleY = (M_PI / 2.0) - (std::atan((pViewport.h / 2.0) / pCamera3D->ViewPlaneDistPixels));
 
     if (uCurrentlyLoadedLevelType == LEVEL_INDOOR) {
-        HalfAngleX = (static_cast<float>(M_PI) / 2.0) - (blv_fov_rad / 2.0);
+        HalfAngleX = (M_PI / 2.0) - (blv_fov_rad / 2.0);
     }
 
     glm::vec3 PlaneVec(0);
